requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
requests>=2.31.0
langchain-groq>=0.1.0
plotly>=5.18.0
pandas>=2.2.0
pdfkit>=1.0.0
jinja2>=3.1.0
wkhtmltopdf>=0.2 